Shelly: Allow choosing between unicast and multicast CoIoT
UDP multicast isn't properly working on some network setups
so we'll default to unicast by configuring our own IP to the
shelly. This has the downside that the Shelly device won't work
any more with other software or multiple nymea setups, so we'll
still allow choosing the multicast option in the setup params.